requirements.txt
python-pptx==0.6.23
Pillow==10.4.0
PyMySQL==1.1.1
requests==2.32.3
python-dotenv==1.0.1